move to own file and add some tests
